+++ fj-doc-base/src/test/java/test/org/fugerit/java/doc/base/facade/TestDocHandlerFacade.java
@@ -3,12 +3,16 @@ package test.org.fugerit.java.doc.base.facade;
 import java.io.ByteArrayOutputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
+import java.nio.charset.StandardCharsets;
 
 import org.fugerit.java.core.cfg.ConfigException;
 import org.fugerit.java.core.cfg.ConfigRuntimeException;
+import org.fugerit.java.core.function.SafeFunction;
 import org.fugerit.java.core.lang.helpers.ClassHelper;
+import org.fugerit.java.doc.base.config.DocConfig;
 import org.fugerit.java.doc.base.config.DocOutput;
 import org.fugerit.java.doc.base.config.DocTypeHandler;
+import org.fugerit.java.doc.base.config.DocTypeHandlerDefault;
 import org.fugerit.java.doc.base.facade.DocHandlerFacade;
 import org.fugerit.java.doc.base.typehandler.markdown.SimpleMarkdownBasicTypeHandler;
 import org.junit.Assert;
@@ -55,5 +59,28 @@ public class TestDocHandlerFacade {
 		}
 		Assert.assertFalse( facade.listHandlersForType( SimpleMarkdownBasicTypeHandler.HANDLER.getType() ).isEmpty() );
 	}
+	
+	@Test
+	public void testFullMap() {
+		DocHandlerFacade facade = new DocHandlerFacade();
+		SafeFunction.apply( () -> {
+			// for full map testing
+			facade.registerHandlerAndId( "test-1" , new TestDocTypeHandler( DocConfig.TYPE_PDF , DocConfig.FORMAT_PDF_A_1B ), true );
+			facade.registerHandlerAndId( "test-2" , new TestDocTypeHandler( DocConfig.TYPE_PDF , DocConfig.TYPE_PDF ), true );
+			facade.registerHandlerAndId( DocConfig.TYPE_PDF , new TestDocTypeHandler( DocConfig.TYPE_PDF , DocConfig.TYPE_PDF ), true );
+			Assert.assertNotNull( facade.findHandler( DocConfig.FORMAT_PDF_A_1B ) );
+		} );
+		
+	}
+	
+}
+
+class TestDocTypeHandler extends DocTypeHandlerDefault {
+	
+	public TestDocTypeHandler(String type, String format ) {
+		super(type, "test", "text/html", StandardCharsets.UTF_8, format);
+	}
+
+	private static final long serialVersionUID = -2844271991860539644L;
 	
 }